cast [goreleaser]
