+++ ethergo/util/export_test.go
@@ -1,6 +1,9 @@
 package util
 
-import "math/big"
+import (
+	"github.com/ethereum/go-ethereum/common"
+	"math/big"
+)
 
 func MakeOptions(options ...CopyOption) TestCopyOptions {
 	return makeOptions(options...)
@@ -37,3 +40,36 @@ func (c copyOptions) GasTipCap() *big.Int {
 func (c copyOptions) TxType() *uint8 {
 	return c.txType
 }
+
+// NullFieldAttribute is a constant used to test the null field attribute.
+// it exports the underlying constant for testing.
+const NullFieldAttribute = nullFieldAttribute
+
+func AddressPtrToString(address *common.Address) string {
+	return addressPtrToString(address)
+}
+
+const (
+	// HashAttr exports hashAttr for testing.
+	HashAttr = hashAttr
+	// FromAttr exports fromAttr for testing.
+	FromAttr = fromAttr
+	// ToAttr exports toAttr for testing.
+	ToAttr = toAttr
+	// DataAttr exports dataAttr for testing.
+	DataAttr = dataAttr
+	// ValueAttr exports valueAttr for testing.
+	ValueAttr = valueAttr
+	// NonceAttr exports nonceAttr for testing.
+	NonceAttr = nonceAttr
+	// GasLimitAttr exports gasLimitAttr for testing.
+	GasLimitAttr = gasLimitAttr
+	// ChainIDAttr exports chainIDAttr for testing.
+	ChainIDAttr = chainIDAttr
+	// GasPriceAttr exports gasPriceAttr for testing.
+	GasPriceAttr = gasPriceAttr
+	// GasFeeCapAttr exports gasFeeCapAttr for testing.
+	GasFeeCapAttr = gasFeeCapAttr
+	// GasTipCapAttr exports gasTipCapAttr for testing.
+	GasTipCapAttr = gasTipCapAttr
+)
